[not verified] Inline util.js inside widget-visibility/editor/index.jsx
